Implemented the mass action version of the z2 x2 - feedback by introducing a complex Cx2z2.
This required the introduction of several different observation functions, where the complex was added to.

Also removed the old way of setting the parameters for the fit setting and improved the data generation function to force the molecules to their steady states by adding more data points at the steady state with decreasing sigma for later timepoints.